+++ crates/runtime/src/value/val_union.rs
@@ -188,9 +188,9 @@ impl ValueRef {
                                     }
                                     let index = must_normalize_index(index, origin_value.len());
                                     if v.is_undefined() {
-                                        origin_value.list_remove_at(index as usize);
+                                        origin_value.list_remove_at(index);
                                     } else {
-                                        origin_value.list_must_set(index as usize, v);
+                                        origin_value.list_must_set(index, v);
                                     }
                                 } else {
                                     panic!(
